Remove slurm plugin from old scaling architecture code
* Remove slurm plugin, utility methods, and unit tests that are no longer used in old scaling architecture
