--- tests/common/schedulers/test_slurm_commands/test_get_jobs_info/squeue_output_empty.txt
@@ -1,1 +1,0 @@
-JOBID                                                                                                                                                                                                   ST                                                                                                                                                                                                      NODES                                                                                                                                                                                                   CPUS                                                                                                                                                                                                    TASKS                                                                                                                                                                                                   MIN_CPUS                                                                                                                                                                                                REASON                                                                                                                                                                                                  TRES_PER_JOB                                                                                                                                                                                            TRES_PER_TASK
