--- tests/common/schedulers/test_slurm_commands/test_get_jobs_info/squeue_output_extra_column.txt
@@ -1,2 +1,0 @@
-JOBID                                                                                                                                                                                                   EXTRA                                                                                                                                                                                                   ST                                                                                                                                                                                                      NODES                                                                                                                                                                                                   CPUS                                                                                                                                                                                                    TASKS                                                                                                                                                                                                   CPUS_PER_TASK                                                                                                                                                                                           MIN_CPUS                                                                                                                                                                                                REASON                                                                                                                                                                                                  TRES_PER_JOB                                                                                                                                                                                            TRES_PER_TASK
-72                                                                                                                                                                                                      72                                                                                                                                                                                                      PD                                                                                                                                                                                                      2                                                                                                                                                                                                       5                                                                                                                                                                                                       5                                                                                                                                                                                                       1                                                                                                                                                                                                       1                                                                                                                                                                                                       Resources                                                                                                                                                                                               N/A                                                                                                                                                                                                     N/A
