--- tests/common/schedulers/test_slurm_commands/test_get_jobs_info/squeue_output_missing_column.txt
@@ -1,2 +1,0 @@
-JOBID                                                                                                                                                                                                   NODES                                                                                                                                                                                                   CPUS                                                                                                                                                                                                    TASKS                                                                                                                                                                                                   CPUS_PER_TASK                                                                                                                                                                                           REASON                                                                                                                                                                                                  TRES_PER_JOB                                                                                                                                                                                            TRES_PER_TASK
-72                                                                                                                                                                                                      2                                                                                                                                                                                                       5                                                                                                                                                                                                       5                                                                                                                                                                                                       1                                                                                                                                                                                                       Resources                                                                                                                                                                                               N/A                                                                                                                                                                                                     N/A
